Beef up packaging scripts, add some documentation

+++ basebox/package-vbox.sh
@@ -14,34 +14,27 @@ vmname=$1
 
 [[ ! -z "$2" ]] && boxbase=$2 || boxbase=$vmname
 
-clear_backup=false
-
 boxfile=$boxbase.box
 
-if [ -f $boxfile ]; then
-  if [ -f $boxfile.old ]; then
-    rm $boxfile.old
-  fi
+if [ -f "$boxfile" ]; then
+  [ -f "$boxfile.old" ] && rm $boxfile.old
   mv $boxfile $boxfile".old"
-  clear_backup=true
 fi
 
-echo "Exporting VM '$vmname' to file '$boxfile' and adding to vagrant as '$boxbase'..."
+echo "Exporting VM '$vmname' to file '$boxfile' ..."
 
 vagrant package --base $vmname --output $boxfile $boxbase
 error=$?
 
-if [[ $error != 0 ]]; then
+if [ $error != 0 ]
+then
   echo "==> $1: Packaging failure!"
-  if [[ "$clear_backup" = true ]]; then
+  if [ -f "$boxfile.old" ]
+  then
     echo "==> $1: Restoring previous version"
-    mv $boxfile.old $boxfile
+    mv "$boxfile.old" "$boxfile"
   fi
   exit $error
 fi
 
-error=$?
-
-if [[ "$clear_backup" = true ]]; then
-  rm $boxfile.old
-fi
+[ -f "$boxfile.old" ] && rm $boxfile.old
